@@ -1,21 +1,36 @@
 {
-  "name": "mnvx/lowrapper",
-  "license": "MIT",
-  "type": "library",
-  "description": "PHP wrapper over LibreOffice converter",
-  "keywords": ["libreoffice", "converter", "word", "pdf", "excel", "docx", "xls", "xlsx"],
-  "homepage": "https://github.com/mnvx/lowrapper",
-  "autoload": {
-    "psr-4": {
-      "Mnvx\\Lowrapper\\": "src/"
+    "name": "darkdarin/lowrapper",
+    "license": "MIT",
+    "type": "library",
+    "description": "PHP wrapper over LibreOffice converter",
+    "keywords": [
+        "libreoffice",
+        "converter",
+        "word",
+        "pdf",
+        "excel",
+        "docx",
+        "xls",
+        "xlsx"
+    ],
+    "homepage": "https://github.com/darkdarin/lowrapper",
+    "autoload": {
+        "psr-4": {
+            "DarkDarin\\Lowrapper\\": "src/"
+        }
+    },
+    "require": {
+        "php": "8.2.*|8.3.*",
+        "symfony/process": "^6.0|^7.0",
+        "psr/log": "^1|^2|^3"
+    },
+    "require-dev": {
+        "phpunit/phpunit": "^10.0",
+        "laravel/pint": "^1.10",
+        "vimeo/psalm": "^5.23"
+    },
+    "scripts": {
+        "pint": "vendor/bin/pint",
+        "psalm": "vendor/bin/psalm --output-format=phpstorm --no-cache"
     }
-  },
-  "require": {
-    "php": "^7.2|^8.0",
-    "symfony/process": "^3.1|^4.1|^5.0|^6.0",
-    "psr/log": "^1|^2|^3"
-  },
-  "require-dev": {
-    "phpunit/phpunit": "^8.0"
-  }
 }
