@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 require __DIR__ . '/../vendor/psr/log/Psr/Log/AbstractLogger.php';
 
 use PSR\Log\AbstractLogger;
@@ -9,7 +11,7 @@ class Logger extends AbstractLogger
     /**
      * @inheritdoc
      */
-    public function log($level, $message, array $context = array())
+    public function log($level, $message, array $context = array()): void
     {
         echo $level . ': <pre>' . $message . '</pre><br/>';
     }
