@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 require __DIR__ . '/../vendor/autoload.php';
 
 ini_set('display_errors', 1);
@@ -7,24 +9,22 @@ ini_set('display_startup_errors', 1);
 error_reporting(E_ALL);
 
 
-
-
 // Example start
 
-use Mnvx\Lowrapper\Converter;
-//use Mnvx\Lowrapper\DocumentType;
-use Mnvx\Lowrapper\LowrapperParameters;
-use Mnvx\Lowrapper\Format;
+use DarkDarin\Lowrapper\Converter;
+use DarkDarin\Lowrapper\Format\TextFormatEnum;
+use DarkDarin\Lowrapper\LowrapperParameters;
+
+//use DarkDarin\Lowrapper\DocumentTypeEnum;
 
 $outputFile = __DIR__ . '/output/html-to-text.text';
 
 $converter = new Converter();
-$parameters = (new LowrapperParameters())
-//    ->setDocumentType(DocumentType::WRITER)
+$parameters = (new LowrapperParameters(TextFormatEnum::TEXT))
+//    ->setDocumentType(DocumentTypeEnum::WRITER)
     ->setInputFile(__DIR__ . '/data/html.html')
 //    ->addOutputFilter('Text (encoded)')
 //    ->addOutputFilter('UTF8')
-    ->setOutputFormat(Format::TEXT_TEXT)
     ->setOutputFile($outputFile);
 
 $converter->convert($parameters);
