@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 require __DIR__ . '/../vendor/autoload.php';
 
 ini_set('display_errors', 1);
@@ -11,18 +13,17 @@ require __DIR__ . '/Logger.php';
 
 // Example start
 
-use Mnvx\Lowrapper\Converter;
-use Mnvx\Lowrapper\LowrapperParameters;
-use Mnvx\Lowrapper\Format;
+use DarkDarin\Lowrapper\Converter;
+use DarkDarin\Lowrapper\Format\TextFormatEnum;
+use DarkDarin\Lowrapper\LowrapperParameters;
 
 $source = file_get_contents(__DIR__ . '/data/html.html');
 
 $converter = new Converter();
 $converter->setLogger(new Logger());
 
-$parameters = (new LowrapperParameters())
-    ->setInputData($source)
-    ->setOutputFormat(Format::TEXT_TEXT);
+$parameters = (new LowrapperParameters(TextFormatEnum::TEXT))
+    ->setInputData($source);
 
 $result = $converter->convert($parameters);
 
