@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 require __DIR__ . '/../vendor/autoload.php';
 
 ini_set('display_errors', 1);
@@ -11,21 +13,20 @@ require __DIR__ . '/Logger.php';
 
 // Example start
 
-use Mnvx\Lowrapper\Converter;
-use Mnvx\Lowrapper\DocumentType;
-use Mnvx\Lowrapper\LowrapperParameters;
-use Mnvx\Lowrapper\Format;
+use DarkDarin\Lowrapper\Converter;
+use DarkDarin\Lowrapper\DocumentTypeEnum;
+use DarkDarin\Lowrapper\Format\TextFormatEnum;
+use DarkDarin\Lowrapper\LowrapperParameters;
 
 $source = file_get_contents(__DIR__ . '/data/html.html');
 
 $converter = new Converter();
 $converter->setLogger(new Logger());
 
-$parameters = (new LowrapperParameters())
-    ->setDocumentType(DocumentType::WRITER)
+$parameters = (new LowrapperParameters(TextFormatEnum::PDF))
+    ->setDocumentType(DocumentTypeEnum::WRITER)
     ->setInputData($source)
-    ->setOutputFile(__DIR__ . '/output/html-to-pdf.pdf')
-    ->setOutputFormat(Format::TEXT_PDF);
+    ->setOutputFile(__DIR__ . '/output/html-to-pdf.pdf');
 
 $converter->convert($parameters);
 
