@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 require __DIR__ . '/../vendor/autoload.php';
 
 ini_set('display_errors', 1);
@@ -7,20 +9,17 @@ ini_set('display_startup_errors', 1);
 error_reporting(E_ALL);
 
 
-
-
 // Example start
 
-use Mnvx\Lowrapper\Converter;
-use Mnvx\Lowrapper\LowrapperParameters;
-use Mnvx\Lowrapper\Format;
+use DarkDarin\Lowrapper\Converter;
+use DarkDarin\Lowrapper\Format\WebFormatEnum;
+use DarkDarin\Lowrapper\LowrapperParameters;
 
 $outputFile = __DIR__ . '/output/odt-to-html.html';
 
 $converter = new Converter();
-$parameters = (new LowrapperParameters())
-    ->setInputFile(__DIR__ . '/data/odt.odt')
-    ->setOutputFormat(Format::WEB_HTML);
+$parameters = (new LowrapperParameters(WebFormatEnum::HTML))
+    ->setInputFile(__DIR__ . '/data/odt.odt');
 
 $result = $converter->convert($parameters);
 
