@@ -1,14 +1,19 @@
 <?php
 
-namespace Mnvx\Lowrapper;
+declare(strict_types=1);
 
+namespace DarkDarin\Lowrapper;
+
+/**
+ * @api
+ */
 interface ConverterInterface
 {
     /**
      * Convert formats
      * @param LowrapperParameters $parameters
-     * @return mixed
+     * @return string|null
      * @throws LowrapperException
      */
-    public function convert(LowrapperParameters $parameters);
+    public function convert(LowrapperParameters $parameters): ?string;
 }
